🔧 Refactor toc/index.ts into multiple files
