+++ src/sync/init.ts
@@ -9,7 +9,8 @@ import { LogLevel } from '../logging';
 import { MyUser } from '../models';
 import { ISession } from '../session/types';
 import { selectors } from '../store';
-import { findProjectsOnPath, loadProjectFromDisk } from '../toc';
+import { loadProjectFromDisk } from '../toc';
+import { findProjectsOnPath } from '../toc/utils';
 import { writeFileToFolder } from '../utils';
 import { startServer } from '../web';
 import { interactiveCloneQuestions } from './clone';
